Keyboard bug fixed & variables from stealth passed to frenzy
The keyboard bug is fixed by changing the system, using overlap. When you overlap with the computer, you press the space bar in order to switch scenes. The scene will have a health point equivalent to the computer. Once you click on the screen 5 times (reducing the health points by 1 each time you do so), you come back to stealth and the keyboard input is then resumed.