Localized all variables + Added camera zoom + Level Design + bug fixes
The camera zooms onto the player by the scale of 4.  The level design is set in an array. Each wall is an object with X, Y-Coordinates, graphic key, and the scale of the graphic. The computers is an array of locations of the computers. There was a bug where after getting caught by the guard, if resetting the scene and the player presses the opposite direction that was last pressed before getting caught, then the player would endlessly move towards that direction indefinitely but fixed it inside of the update function.