CLEAR CONDITION UPDATED AND BUG FIXES
I realized that sometimes, even if the target's value goes to exactly zero, the computational process in the computer does not fully match up zero. So to prevenet that, I just made less than or equal to. I also made clear condition so that the exit point does not interact with the player until the number of targets is zero.